Реализована инъекция зависимостей с использованием Ninject.
* Добавлены NuGet-пакеты для интеграции Ninject.
* Откатил пакеты Microsoft.AspNet.WebApi.Client, Microsoft.AspNet.WebApi.Core, Microsoft.AspNet.WebApi.Owin и Newtonsoft.Json на более старые версии, чтобы обеспечить совместимость с пакетами интеграции Ninject без байндинг-редиректов.
* Выделен интерфейс IMongoDBStorage, поля Revisions и Users сделаны свойствами.
* Конструктор CounterController сделан публичным и принимает параметр IMongoDBStorage для инъекции.
* Реализация IAuthenticationFilter вынесена из BasicAuthenticationAttribute в класс BasicAuthenticationFilter. Конструктор BasicAuthenticationFilter сделан публичным и принимает параметр IMongoDBStorage для инъекции.
* Удалён класс MongoDBStorageFactory.
* Байндинг зависимостей производится в классе Startup.
